# api-backend/openapi.yaml
openapi: 3.0.0
info:
  title: MeusFilmes API
  version: 1.0.0
  description: API for the MeusFilmes application, providing access to movie data, user lists, and watched history.
servers:
  - url: /api # Adjusted to reflect Nginx proxy
    description: Main API access via Nginx proxy
paths:
...
  /lists/:
    get:
      summary: Lista todas as listas do usuário autenticado
      tags:
        - Lists
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Listas do usuário
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    is_main:
                      type: boolean
    post:
      summary: Cria uma nova lista personalizada para o usuário autenticado
      tags:
        - Lists
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        '201':
          description: Lista criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  is_main:
                    type: boolean
                  user_id:
                    type: integer
        '400':
          description: Nome ausente
  /lists/add_movie:
    post:
      summary: Adiciona um filme a uma lista do usuário
      tags:
        - Lists
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                list_id:
                  type: integer
                tmdb_id:
                  type: integer
              required:
                - list_id
                - tmdb_id
      responses:
        '201':
          description: Filme adicionado à lista
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  tmdb_id:
                    type: integer
        '400':
          description: Dados ausentes
        '404':
          description: Lista não encontrada
        '409':
          description: Filme já está na lista
        '500':
          description: Erro ao salvar filme
  /lists/{list_id}:
    get:
      summary: Detalhes de uma lista do usuário
      tags:
        - Lists
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: list_id
          schema:
            type: integer
          required: true
          description: ID da lista
      responses:
        '200':
          description: Detalhes da lista
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  is_main:
                    type: boolean
                  user_id:
                    type: integer
                  movies:
                    type: array
                    items:
                      type: object
                      properties:
                        tmdb_id:
                          type: integer
        '404':
          description: Lista não encontrada
    delete:
      summary: Exclui uma lista personalizada do usuário
      tags:
        - Lists
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: list_id
          schema:
            type: integer
          required: true
          description: ID da lista
      responses:
        '200':
          description: Lista excluída
        '404':
          description: Lista não encontrada
        '403':
          description: Não é permitido excluir listas principais
    put:
      summary: Renomeia uma lista do usuário
      tags:
        - Lists
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: list_id
          schema:
            type: integer
          required: true
          description: ID da lista
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        '200':
          description: Lista renomeada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  is_main:
                    type: boolean
                  user_id:
                    type: integer
        '400':
          description: Nome ausente
        '404':
          description: Lista não encontrada
        '403':
          description: Não é permitido renomear listas principais
  /lists/{list_id}/movies/{tmdb_id}:
    delete:
      summary: Remove um filme de uma lista do usuário
      tags:
        - Lists
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: list_id
          schema:
            type: integer
          required: true
          description: ID da lista
        - in: path
          name: tmdb_id
          schema:
            type: integer
          required: true
          description: ID do filme no TMDB
      responses:
        '200':
          description: Filme removido da lista
        '404':
          description: Lista ou filme não encontrado
  /lists/public/{list_id}:
    get:
      summary: Get public details of a list by its ID (no authentication required)
      tags:
        - Lists
      parameters:
        - in: path
          name: list_id
          schema:
            type: integer
          required: true
          description: List ID
      responses:
        '200':
          description: Public list details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  is_main:
                    type: boolean
                  user_id:
                    type: integer
                  movies:
                    type: array
                    items:
                      type: object
                      properties:
                        tmdb_id:
                          type: integer
        '404':
          description: List not found
  /auth/register:
    post:
      summary: Registra um novo usuário
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: João da Silva
                email:
                  type: string
                  format: email
                  example: joao@email.com
                password:
                  type: string
                  format: password
                  example: senha123
              required:
                - name
                - email
                - password
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      email:
                        type: string
        '400':
          description: Campos obrigatórios ausentes
        '409':
          description: Email já registrado
  /auth/login:
    post:
      summary: Login do usuário
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: joao@email.com
                password:
                  type: string
                  format: password
                  example: senha123
              required:
                - email
                - password
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      email:
                        type: string
        '400':
          description: Campos obrigatórios ausentes
        '401':
          description: Credenciais inválidas
  /auth/recovery:
    post:
      summary: Solicita recuperação de senha
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: joao@email.com
              required:
                - email
      responses:
        '200':
          description: Instruções de recuperação enviadas (ou resposta genérica)
        '400':
          description: Email ausente
        '404':
          description: Email não encontrado
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []

# You will need to manually add definitions for all your other API endpoints
# (e.g., /login, /lists, /movie/watched, /tmdb/*, etc.)
# following the OpenAPI 3.0 specification format.
# Pay close attention to request bodies, parameters (path, query), and responses for each.